# smart_home/smart home/backend/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "智能家居综合管理系统后端",
    "description": "智能家居综合管理系统 Python FastAPI 后端接口服务。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/homes/{home_id}": {
      "get": {
        "tags": [
          "家庭空间"
        ],
        "summary": "Get Home",
        "operationId": "get_home_api_homes__home_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "家庭空间"
        ],
        "summary": "Update Home",
        "operationId": "update_home_api_homes__home_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "家庭空间"
        ],
        "summary": "Delete Home",
        "operationId": "delete_home_api_homes__home_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/homes/{home_id}/scenes": {
      "post": {
        "tags": [
          "情景模式"
        ],
        "summary": "新增情景模式",
        "operationId": "create_scene_api_homes__home_id__scenes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Scene Api Homes  Home Id  Scenes Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "情景模式"
        ],
        "summary": "查询情景模式列表",
        "operationId": "list_scenes_api_homes__home_id__scenes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Scenes Api Homes  Home Id  Scenes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenes/{scene_id}": {
      "put": {
        "tags": [
          "情景模式"
        ],
        "summary": "修改情景模式",
        "operationId": "update_scene_api_scenes__scene_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scene Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SceneUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Scene Api Scenes  Scene Id  Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "情景模式"
        ],
        "summary": "删除情景模式",
        "operationId": "delete_scene_api_scenes__scene_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Scene Api Scenes  Scene Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scenes/{scene_id}/execute": {
      "post": {
        "tags": [
          "情景模式"
        ],
        "summary": "执行情景模式",
        "operationId": "execute_scene_api_scenes__scene_id__execute_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "scene_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Scene Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Execute Scene Api Scenes  Scene Id  Execute Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/homes/{home_id}/system/config": {
      "get": {
        "tags": [
          "系统配置与日志"
        ],
        "summary": "Get Config",
        "operationId": "get_config_api_homes__home_id__system_config_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "系统配置与日志"
        ],
        "summary": "Update Config",
        "operationId": "update_config_api_homes__home_id__system_config_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemConfigUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/homes/{home_id}/system/logs": {
      "get": {
        "tags": [
          "系统配置与日志"
        ],
        "summary": "List Logs",
        "operationId": "list_logs_api_homes__home_id__system_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "home_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Home Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HomeUpdateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          }
        },
        "type": "object",
        "title": "HomeUpdateRequest"
      },
...
      "SceneAction": {
        "properties": {
          "device_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Device Id",
            "description": "目标设备编号"
          },
          "target_state": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Target State",
            "description": "目标状态"
          },
          "param_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Param Value",
            "description": "可选参数"
          },
          "sort_no": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Sort No",
            "description": "执行顺序",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "device_id",
          "target_state"
        ],
        "title": "SceneAction"
      },
      "SceneCreate": {
        "properties": {
          "scene_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Scene Name",
            "description": "情景名称"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "description": "是否启用",
            "default": true
          },
          "actions": {
            "items": {
              "$ref": "#/components/schemas/SceneAction"
            },
            "type": "array",
            "title": "Actions",
            "description": "情景动作"
          }
        },
        "type": "object",
        "required": [
          "scene_name"
        ],
        "title": "SceneCreate"
      },
      "SceneUpdate": {
        "properties": {
          "scene_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Scene Name",
            "description": "情景名称"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled",
            "description": "是否启用"
          },
          "actions": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SceneAction"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actions",
            "description": "情景动作"
          }
        },
        "type": "object",
        "title": "SceneUpdate"
      },
...
      "SystemConfigUpdateRequest": {
        "properties": {
          "alarm_smoke_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alarm Smoke Threshold"
          },
          "alarm_gas_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alarm Gas Threshold"
          },
          "temperature_high_threshold": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature High Threshold"
          },
          "auto_alarm_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Alarm Enabled"
          },
          "simulation_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Simulation Enabled"
          }
        },
        "type": "object",
        "title": "SystemConfigUpdateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}